♻️ Move `formatXAgo` to `_pub_shared`
